4_professional_backend/day28_capstone_project/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
sqlalchemy[asyncio]
aiosqlite==0.18.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.5
httpx==0.24.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0
aiofiles==23.1.0
pydantic[email]
python-multipart
argon2-cffi